split some functions into other scripts to make main script shorter, add readme
